requirements.txt
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
countryinfo==0.1.2
Flask==3.0.3
Flask-Session==0.8.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lb-matching-tools==2024.1.30.1
liblistenbrainz==0.5.5
MarkupSafe==3.0.1
more-itertools==10.5.0
msgspec==0.18.6
mutagen==1.46.0
nmslib-metabrainz==2.1.3
numpy==2.1.2
packaging==24.1
peewee==3.17.7
psutil==6.1.0
psycopg2-binary==2.9.10
py-sonic==1.0.1
pybind11==2.13.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.1
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
spotipy==2.24.0
threadpoolctl==3.5.0
tqdm==4.66.5
troi==2024.8.30.1
types-peewee==3.17.7.20241017
ujson==5.10.0
Unidecode==1.3.8
urllib3==2.2.3
Werkzeug==3.0.4
